refactor kpi-page + % display
