+++ KPI-team1-frontend/src/pages/KpiPage.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import { supabase } from "../supabase";
 import { KpiExtended } from "../model/kpi";
-import KpiDetailModalPage from "./KpiDetailModalPage";
+import KpiDetailModalPage from "./KpiModalPage/KpiDetailModalPage";
 import { getDisplayValueByPeriodicity } from "../helpers/kpiHelpers";
 import { useOutletContext, useParams } from "react-router-dom";
 import { Circles } from "../model/circle";
@@ -24,14 +24,20 @@ const HEADER_KPI_COLUMNS: GridColDef[] = [
     sortable: true,
     hideable: false,
     headerAlign: "center",
-  },
-  {
-    headerName: "Target",
-    field: "kpi_target",
-    width: 150,
-    sortable: false,
-    headerAlign: "center",
-    align: "center",
+    renderCell: (params) => {
+      if (params.row.unit === "%") {
+        return (
+          <>
+            <span>{params.value}</span>
+            <div className="ml-4 w-6 h-6 bg-amber-400 rounded-[100px] flex-col justify-center items-center gap-2 inline-flex">
+              <div className="w-[19px] text-center text-zinc-700 text-base font-medium font-['Inter']">
+                %
+              </div>
+            </div>
+          </>
+        );
+      }
+    },
   },
   {
     headerName: "Latest Value",
@@ -39,8 +45,17 @@ const HEADER_KPI_COLUMNS: GridColDef[] = [
     width: 150,
     sortable: true,
     headerAlign: "center",
-
     align: "center",
+    renderCell: (params) => {
+      const value = params.value as number;
+      if (value === null) {
+        return null;
+      } else if (params.row.unit === "%") {
+        return value.toFixed(2) + "%";
+      } else {
+        return value.toLocaleString("fr-ch");
+      }
+    },
   },
   {
     headerName: "Last Update",
@@ -65,6 +80,14 @@ const HEADER_KPI_COLUMNS: GridColDef[] = [
     width: 300,
     sortable: false,
     filterable: false,
+    headerAlign: "center",
+    align: "center",
+  },
+  {
+    headerName: "Target",
+    field: "kpi_target",
+    width: 150,
+    sortable: false,
     headerAlign: "center",
     align: "center",
   },
